+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Change function `Parser.parse` to return Async Result
+- [**BC**] Change `Resolver.resolve` function to return a `ResolveError` on error
+- Add `Resolver.resolveAsync` function
+- [_Internal_] Fix `asyncResult` computation expression
 
 ## 2.1.0 - 2021-08-11
 - Update dependencies